Resolve 9 merge conflicts adapting to development conventions:
- Update builders to use factory methods (From, Create, New)
- Rename *Id test files to *Identifier (ProjectId→ProjectIdentifier, etc.)
- Use no-underscore field naming in test fixtures
- Adopt development's ValidationBehavior approach (real validators)
- Keep both new architecture tests and development's IValueObject test
- Replace Task.Delay with TaskCompletionSource in SyncSchedulerServiceTests

All 693 tests pass (0 failures).

+++ src/backend/Services/Documentation/BauDoku.Documentation.Application/Queries/Dtos/InstallationDto.cs
@@ -19,7 +19,7 @@ public sealed record InstallationDto(
     string QualityGrade,
     string? Description,
     string? CableType,
-    int? CrossSection,
+    decimal? CrossSection,
     string? CableColor,
     int? ConductorCount,
     int? DepthMm,
